+++ core/src/main/kotlin/com/malinskiy/marathon/execution/DevicePoolMessage.kt
@@ -5,7 +5,7 @@ import com.malinskiy.marathon.device.Device
 sealed class DevicePoolMessage {
     data class AddDevice(val device: Device) : DevicePoolMessage()
     data class RemoveDevice(val device: Device) : DevicePoolMessage()
-    data class TestExecutionFinished(val device: Device, val sender: DeviceAktor) : DevicePoolMessage()
-    data class Ready(val device: Device, val sender: DeviceAktor) : DevicePoolMessage()
+    data class TestExecutionFinished(val device: Device, val sender: DeviceActor) : DevicePoolMessage()
+    data class Ready(val device: Device, val sender: DeviceActor) : DevicePoolMessage()
     object Terminate : DevicePoolMessage()
 }
